Adding the running of the queries and additions to provenance system
Main changes are to concatenate the queries to be run by one function. One query
will generate a bigquery table and the rest all produce a csv file from a dataframe.

Additions to sql.py to provide convenience functions for those and to provndoc_utils.py to add some functionality to the DAGs to detect whether the output of a given query is supposed to be a table or a file.